test: increase test coverage and refactor inline tests
- Extracted 16 crowded inline tests from `src/protocol/ptp/node.rs` into `src/protocol/ptp/node/tests_unit.rs`
- Added unit tests for `PlaybackController` rate control (fast forward and rewind)
- Added unit tests for `RaopRtspSession` Options response parsing to cover `Apple-Response` extraction
- Ran clippy and formatter, successfully checked and passed all unit and integration tests
