+++ src/control/tests/playback.rs
@@ -131,3 +131,24 @@ async fn test_playback_controller_set_shuffle_and_repeat() {
     // it returns an error, state might remain unchanged.
     assert!(res.is_err());
 }
+
+#[tokio::test]
+async fn test_fast_forward_and_rewind() {
+    let config = AirPlayConfig::default();
+    let manager = Arc::new(ConnectionManager::new(config));
+    let controller = crate::control::playback::PlaybackController::new(manager);
+
+    // Initial position should be 0.0
+    assert_eq!(controller.state().await.position_secs, 0.0);
+
+    // Fast forward should attempt to skip forward 10s. Since it's disconnected, it'll fail,
+    // but the state doesn't update on failure. We can mock it or just verify it fails correctly.
+    let result = controller.fast_forward().await;
+    assert!(
+        result.is_err(),
+        "fast_forward should fail when disconnected"
+    );
+
+    let result = controller.rewind().await;
+    assert!(result.is_err(), "rewind should fail when disconnected");
+}
