+++ src/protocol/ptp/node.rs
@@ -1399,285 +1399,4 @@ pub fn create_client_node(
 /// `check_announce_timeout`). These tests live inside the module so they
 /// can access private fields and methods without making them pub.
 #[cfg(test)]
-mod tests_unit {
-    use std::net::SocketAddr;
-    use std::str::FromStr;
-    use std::sync::Arc;
-    use std::time::Duration;
-
-    use super::{EffectiveRole, PtpNode, PtpNodeConfig};
-    use crate::protocol::ptp::clock::PtpRole;
-    use crate::protocol::ptp::handler::create_shared_clock;
-
-    /// Build a minimal `PtpNode` bound to an ephemeral loopback port.
-    async fn make_node(our_priority1: u8, our_clock_id: u64) -> PtpNode {
-        let sock = Arc::new(tokio::net::UdpSocket::bind("127.0.0.1:0").await.unwrap());
-        let clock = create_shared_clock(our_clock_id, PtpRole::Master);
-        let config = PtpNodeConfig {
-            clock_id: our_clock_id,
-            priority1: our_priority1,
-            priority2: 128,
-            ..Default::default()
-        };
-        PtpNode::new(sock, None, clock, config)
-    }
-
-    // ── compare_priority ─────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn test_compare_priority_remote_wins_lower_p1() {
-        // We have p1=255 (worst possible), remote has p1=128 → remote is better.
-        let node = make_node(255, 0xAAAA).await;
-        assert!(
-            node.compare_priority(128, 128, 0xBBBB),
-            "Remote with lower priority1 must win"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_compare_priority_we_win_with_lower_p1() {
-        // We have p1=64, remote has p1=128 → we are better.
-        let node = make_node(64, 0xAAAA).await;
-        assert!(
-            !node.compare_priority(128, 128, 0xBBBB),
-            "Remote with higher priority1 must NOT win"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_compare_priority_equal_p1_remote_wins_lower_p2() {
-        // Both p1=128. Remote p2=64 < our p2=128 → remote wins on priority2.
-        let node = make_node(128, 0xAAAA).await;
-        assert!(
-            node.compare_priority(128, 64, 0xBBBB),
-            "Remote with lower priority2 (tie on p1) must win"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_compare_priority_equal_p1_we_win_higher_remote_p2() {
-        // Both p1=128. Remote p2=200 > our p2=128 → we win.
-        let node = make_node(128, 0xAAAA).await;
-        assert!(
-            !node.compare_priority(128, 200, 0xBBBB),
-            "Remote with higher priority2 (tie on p1) must NOT win"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_compare_priority_tiebreak_on_lower_clock_id() {
-        // Both p1=128, p2=128. Remote clock_id=0x0001 < ours=0xAAAA → remote wins.
-        let node = make_node(128, 0xAAAA).await;
-        assert!(
-            node.compare_priority(128, 128, 0x0001),
-            "Remote with lower clock_id (tie on both priorities) must win"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_compare_priority_tiebreak_on_higher_clock_id_we_win() {
-        // Both p1=128, p2=128. Remote clock_id=0xFFFF > ours=0xAAAA → we win.
-        let node = make_node(128, 0xAAAA).await;
-        assert!(
-            !node.compare_priority(128, 128, 0xFFFF),
-            "Remote with higher clock_id (tie on both priorities) must NOT win"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_compare_priority_identical_parameters_is_false() {
-        // If remote and local have the exact same values, remote does not win
-        // (since `remote_clock_id < self.config.clock_id` is false when equal).
-        let node = make_node(128, 0xAAAA).await;
-        assert!(
-            !node.compare_priority(128, 128, 0xAAAA),
-            "Identical parameters must not trigger a role switch"
-        );
-    }
-
-    // ── process_announce ─────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn test_process_announce_switches_to_slave_when_remote_better() {
-        let mut node = make_node(255, 0xAAAA).await;
-        assert_eq!(node.role, EffectiveRole::Master);
-
-        let src = SocketAddr::from_str("192.168.1.100:320").unwrap();
-        // Remote p1=128 < our p1=255 → remote is better.
-        node.process_announce(0xBBBB_CCCC_DDDD_EEEE, 128, 128, src);
-
-        assert_eq!(
-            node.role,
-            EffectiveRole::Slave,
-            "Should switch to Slave when a better-priority Announce arrives"
-        );
-        assert!(
-            node.remote_master.is_some(),
-            "remote_master must be populated after switching to Slave"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_process_announce_stays_master_when_remote_worse() {
-        let mut node = make_node(64, 0xAAAA).await;
-        assert_eq!(node.role, EffectiveRole::Master);
-
-        let src = SocketAddr::from_str("192.168.1.100:320").unwrap();
-        // Remote p1=128 > our p1=64 → we are better, stay Master.
-        node.process_announce(0xBBBB_CCCC_DDDD_EEEE, 128, 128, src);
-
-        assert_eq!(
-            node.role,
-            EffectiveRole::Master,
-            "Should stay Master when we have better priority"
-        );
-        assert!(
-            node.remote_master.is_none(),
-            "remote_master must remain None when we stay Master"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_process_announce_ignores_own_clock_id() {
-        // Even if the priority would be better, an Announce with our own clock_id
-        // (e.g. a reflected packet) must be silently dropped.
-        let our_clock_id = 0xAAAA_BBBB_CCCC_DDDD;
-        let mut node = make_node(255, our_clock_id).await;
-        let src = SocketAddr::from_str("192.168.1.100:320").unwrap();
-
-        node.process_announce(our_clock_id, 1, 1, src); // perfect priority but our ID
-
-        assert_eq!(
-            node.role,
-            EffectiveRole::Master,
-            "Own clock_id in Announce must be ignored"
-        );
-        assert!(
-            node.remote_master.is_none(),
-            "remote_master must not be set after ignoring own Announce"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_process_announce_updates_last_announce_when_staying_master() {
-        // When we receive an Announce from a known (but worse) remote, the
-        // remote_master record's last_announce must be refreshed (if it exists).
-        let mut node = make_node(64, 0xAAAA).await;
-
-        // First announce that doesn't switch us (remote is worse).
-        let src = SocketAddr::from_str("192.168.1.100:320").unwrap();
-        node.process_announce(0xBBBB, 128, 128, src);
-        // Still master, no remote_master entry.
-        assert!(node.remote_master.is_none());
-
-        // Now manually install a remote_master so we start as Slave.
-        node.role = EffectiveRole::Slave;
-        node.remote_master = Some(super::RemoteMaster {
-            grandmaster_identity: 0xBBBB,
-            priority1: 128,
-            priority2: 128,
-            event_addr: SocketAddr::from_str("192.168.1.100:319").unwrap(),
-            general_addr: src,
-            last_announce: tokio::time::Instant::now(),
-        });
-        // Tweak our priority to make the remote worse so this Announce won't re-trigger slave.
-        node.config.priority1 = 64;
-
-        // Re-process same remote with same clock_id while we have it tracked.
-        node.process_announce(0xBBBB, 128, 128, src);
-        // remote_master entry is still there (we refreshed it).
-        assert!(
-            node.remote_master.is_some(),
-            "remote_master entry must be preserved when remote sends a new Announce"
-        );
-    }
-
-    // ── check_announce_timeout ────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn test_announce_timeout_reverts_to_master() {
-        let mut node = make_node(255, 0xAAAA).await;
-
-        // First become Slave via Announce.
-        let src = SocketAddr::from_str("192.168.1.100:320").unwrap();
-        node.process_announce(0xBBBB, 128, 128, src);
-        assert_eq!(node.role, EffectiveRole::Slave);
-
-        // Set a very short timeout so it triggers immediately.
-        node.announce_timeout = Duration::from_nanos(1);
-
-        // Sleep a tiny amount so last_announce.elapsed() > 1ns.
-        tokio::time::sleep(Duration::from_millis(5)).await;
-
-        node.check_announce_timeout();
-
-        assert_eq!(
-            node.role,
-            EffectiveRole::Master,
-            "Must revert to Master after announce timeout"
-        );
-        assert!(
-            node.remote_master.is_none(),
-            "remote_master must be cleared after timeout"
-        );
-        // All pending slave state must be cleared.
-        assert!(node.pending_t1.is_none(), "pending_t1 must be cleared");
-        assert!(node.pending_t2.is_none(), "pending_t2 must be cleared");
-        assert!(node.pending_t3.is_none(), "pending_t3 must be cleared");
-        assert!(
-            node.delay_req_sent_at.is_none(),
-            "delay_req_sent_at must be cleared"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_announce_timeout_does_not_fire_when_recent() {
-        let mut node = make_node(255, 0xAAAA).await;
-
-        let src = SocketAddr::from_str("192.168.1.100:320").unwrap();
-        node.process_announce(0xBBBB, 128, 128, src);
-        assert_eq!(node.role, EffectiveRole::Slave);
-
-        // Very long timeout — must NOT fire right after the Announce.
-        node.announce_timeout = Duration::from_secs(60);
-        node.check_announce_timeout();
-
-        assert_eq!(
-            node.role,
-            EffectiveRole::Slave,
-            "Must stay Slave when announce has not timed out"
-        );
-        assert!(
-            node.remote_master.is_some(),
-            "remote_master must remain set when announce is recent"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_announce_timeout_is_no_op_when_no_remote_master() {
-        // Already Master with no remote_master — calling check_announce_timeout
-        // must be a no-op and must not panic.
-        let mut node = make_node(128, 0xAAAA).await;
-        assert_eq!(node.role, EffectiveRole::Master);
-        assert!(node.remote_master.is_none());
-
-        node.announce_timeout = Duration::from_nanos(1);
-        tokio::time::sleep(Duration::from_millis(5)).await;
-        node.check_announce_timeout(); // must not panic
-
-        assert_eq!(node.role, EffectiveRole::Master);
-    }
-
-    // ── Delay_Req timeout / retry (DELAY_REQ_TIMEOUT) ────────────────────────
-
-    /// Verify the `DELAY_REQ_TIMEOUT` constant matches the expected 1-second value.
-    /// This value was tuned to balance responsiveness and avoiding spurious retries.
-    #[test]
-    fn test_delay_req_timeout_constant_is_one_second() {
-        assert_eq!(
-            super::DELAY_REQ_TIMEOUT,
-            Duration::from_secs(1),
-            "DELAY_REQ_TIMEOUT must be 1 second"
-        );
-    }
-}
+mod tests_unit;
